# swagger.json
{"swagger":"2.0","info":{"version":"0.0.15","title":"Thingoo REST API","contact":{"name":"Bartlomiej Rasztabiga","url":"https://github.com/BartlomiejRasztabiga","email":"contact@rasztabiga.me"}},"host":"localhost","basePath":"/","tags":[{"name":"device-controller","description":"Device Controller"},{"name":"entity-controller","description":"Entity Controller"},{"name":"reading-controller","description":"Reading Controller"},{"name":"user-controller","description":"User Controller"}],"paths":{"/devices":{"get":{"tags":["device-controller"],"summary":"getAllDevices","operationId":"getAllDevicesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/DeviceDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["device-controller"],"summary":"registerDevice","operationId":"registerDeviceUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"registerDeviceDto","description":"registerDeviceDto","required":true,"schema":{"$ref":"#/definitions/RegisterDeviceDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DeviceDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/devices/{id}":{"get":{"tags":["device-controller"],"summary":"getDeviceById","operationId":"getDeviceByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DeviceDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["device-controller"],"summary":"deleteDeviceById","operationId":"deleteDeviceByIdUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/devices/{id}/entities":{"get":{"tags":["device-controller"],"summary":"getEntitiesForDevice","operationId":"getEntitiesForDeviceUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/EntityDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/entities/{id}":{"get":{"tags":["entity-controller"],"summary":"getEntityById","operationId":"getEntityByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/EntityDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/entities/{id}/readings":{"get":{"tags":["entity-controller"],"summary":"getReadingsForEntity","operationId":"getReadingsForEntityUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ReadingDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/readings":{"get":{"tags":["reading-controller"],"summary":"getReadings","operationId":"getReadingsUsingGET","produces":["*/*"],"parameters":[{"name":"device_key","in":"query","description":"device_key","required":true,"type":"string"},{"name":"entity_key","in":"query","description":"entity_key","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ReadingDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["reading-controller"],"summary":"saveReading","operationId":"saveReadingUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"saveReadingDto","description":"saveReadingDto","required":true,"schema":{"$ref":"#/definitions/SaveReadingDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ReadingDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/users":{"get":{"tags":["user-controller"],"summary":"getAllUsers","operationId":"getAllUsersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/User"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/users/me":{"get":{"tags":["user-controller"],"summary":"getCurrentUser","operationId":"getCurrentUserUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/User"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"securityDefinitions":{"spring_oauth":{"type":"oauth2","authorizationUrl":"https://dev.thingoo.xyz/auth/realms/Thingoo/protocol/openid-connect/auth","tokenUrl":"https://dev.thingoo.xyz/auth/realms/Thingoo/protocol/openid-connect/token","flow":"accessCode"}},"definitions":{"DeviceDto":{"type":"object","required":["entities","id","key","macAddress"],"properties":{"displayName":{"type":"string"},"entities":{"type":"array","items":{"$ref":"#/definitions/EntityDto"}},"id":{"type":"integer","format":"int32"},"key":{"type":"string"},"macAddress":{"type":"string"}},"title":"DeviceDto"},"EntityDto":{"type":"object","required":["id","key","type","unitDisplayName","unitType"],"properties":{"displayName":{"type":"string"},"id":{"type":"integer","format":"int32"},"key":{"type":"string"},"type":{"type":"string","enum":["ACTUATOR","SENSOR"]},"unitDisplayName":{"type":"string"},"unitType":{"type":"string","enum":["BOOLEAN","DECIMAL","INTEGER","STRING"]}},"title":"EntityDto"},"ReadingDto":{"type":"object","required":["id","timestamp","value"],"properties":{"id":{"type":"integer","format":"int32"},"timestamp":{"type":"string","format":"date-time"},"value":{"type":"string"}},"title":"ReadingDto"},"RegisterDeviceDto":{"type":"object","required":["entities","key","macAddress"],"properties":{"displayName":{"type":"string"},"entities":{"type":"array","items":{"$ref":"#/definitions/RegisterEntityDto"}},"key":{"type":"string"},"macAddress":{"type":"string"}},"title":"RegisterDeviceDto"},"RegisterEntityDto":{"type":"object","required":["key","type","unitDisplayName","unitType"],"properties":{"displayName":{"type":"string"},"key":{"type":"string"},"type":{"type":"string","enum":["ACTUATOR","SENSOR"]},"unitDisplayName":{"type":"string"},"unitType":{"type":"string","enum":["BOOLEAN","DECIMAL","INTEGER","STRING"]}},"title":"RegisterEntityDto"},"SaveReadingDto":{"type":"object","required":["deviceKey","entityKey","value"],"properties":{"deviceKey":{"type":"string"},"entityKey":{"type":"string"},"value":{"type":"string"}},"title":"SaveReadingDto"},"User":{"type":"object","required":["email","enabled","firstName","id","lastName","username"],"properties":{"email":{"type":"string"},"enabled":{"type":"boolean"},"firstName":{"type":"string"},"id":{"type":"string","format":"uuid"},"lastName":{"type":"string"},"username":{"type":"string"}},"title":"User"}}}